README: Add Usage and other documentation
The original subject of this commit is:

    Snapshot to be continued tomorrow

It's not tomorrow anymore. The day after I'd made the commit, Peter
McIntyre asked me to help him with something else. That lasted for about
four months. Now I have no recollection of how I wanted to "continue
tomorrow".

I only know that I was making the repo publishable. So commit what I
have and see what more there is to do.