pomdp_play: Make it runnable
I had turned piglet_pbvi.pbvi.run into piglet_pbvi.pbvi.generator, but
not adapted pomdp_play accordingly. Fix this.

Also remove the setting of the matplotlib backend. I'm on a Mac, so
GTK3Agg doesn't work out of the box. But the default works. Now it might
break somewhere else, of course.

Also change the number of expansions of a later loop, so that it has the
same output as NAIVE.

(pomdp_play isn't really intended for other people to use.)